backend/requirements.txt
flask
flask-cors
python-dotenv
sqlalchemy
psycopg2-binary
requests
supabase

